+++ vsphere/8.0/vcsa/inspec/vmware-vcsa-8.0-stig-baseline/envoy/controls/VCRP-80-000073.rb
@@ -3,7 +3,7 @@ control 'VCRP-80-000073' do
   desc  'Writing events to a centralized management audit system offers many benefits to the enterprise over having dispersed logs. Centralized management of audit records and logs provides for efficiency in maintenance and management of records, enterprise analysis of events, and backup and archiving of event records enterprise-wide. The web server and related components are required to be capable of writing logs to centralized audit log servers.'
   desc  'rationale', ''
   desc  'check', "
-    By default there is a vmware-services-rhttpproxy.conf rsyslog configuration file that includes the service logs when syslog is configured on vCenter but it must be verified.
+    By default, there is a vmware-services-rhttpproxy.conf rsyslog configuration file that includes the service logs when syslog is configured on vCenter, but it must be verified.
 
     At the command prompt, run the following command: