+++ vsphere/8.0/vcsa/inspec/vmware-vcsa-8.0-stig-baseline/postgresql/controls/VCPG-80-000038.rb
@@ -1,9 +1,9 @@
 control 'VCPG-80-000038' do
   title 'The vCenter PostgreSQL service must encrypt passwords for user authentication.'
   desc  "
-    The DoD standard for authentication is DoD-approved PKI certificates.
+    The DOD standard for authentication is DOD-approved PKI certificates.
 
-    Authentication based on User ID and Password may be used only when it is not possible to employ a PKI certificate, and requires AO approval.
+    Authentication based on User ID and Password may be used only when it is not possible to employ a PKI certificate and requires AO approval.
 
     In such cases, database passwords stored in clear text, using reversible encryption, or using unsalted hashes would be vulnerable to unauthorized disclosure. Database passwords must always be in the form of one-way, salted hashes when stored internally or externally to the database management system (DBMS).
   "
@@ -15,9 +15,11 @@ control 'VCPG-80-000038' do
 
     Expected result:
 
-    md5
+    scram-sha-256
 
     If the output does not match the expected result, this is a finding.
+
+    Note: Prior to Update 2, \"md5\" is the expected result.
   "
   desc 'fix', "
     A script is included with vCenter to generate a PostgreSQL STIG configuration.
@@ -44,6 +46,6 @@ control 'VCPG-80-000038' do
   sql = postgres_session("#{input('postgres_user')}", "#{input('postgres_pass')}", "#{input('postgres_host')}")
 
   describe sql.query('SHOW password_encryption;', ["#{input('postgres_default_db')}"]) do
-    its('output') { should cmp 'md5' }
+    its('output') { should cmp 'scram-sha-256' }
   end
 end