+++ vsphere/8.0/vcsa/inspec/vmware-vcsa-8.0-stig-baseline/sts/controls/VCST-80-000081.rb
@@ -9,7 +9,7 @@ control 'VCST-80-000081' do
   "
   desc  'rationale', ''
   desc  'check', "
-    By default, a vmware-services-sso-services.conf rsyslog configuration file that includes the service logs when syslog is configured on vCenter, but it must be verified.
+    By default, a vmware-services-sso-services.conf rsyslog configuration file includes the service logs when syslog is configured on vCenter, but it must be verified.
 
     At the command prompt, run the following command: