+++ vsphere/8.0/vsphere/inspec/vmware-vsphere-8.0-stig-baseline/vm/controls/VMCH-80-000189.rb
@@ -17,7 +17,7 @@ control 'VMCH-80-000189' do
 
     If the virtual machine advanced setting \"isolation.tools.copy.disable\" is not set to \"true\", this is a finding.
 
-    If the virtual machine advanced setting \"isolation.tools.copy.disable\" does NOT exist, this is NOT a finding.
+    If the virtual machine advanced setting \"isolation.tools.copy.disable\" does not exist, this is not a finding.
   "
   desc  'fix', "
     For each virtual machine do the following: