+++ vsphere/8.0/vsphere/inspec/vmware-vsphere-8.0-stig-baseline/CHANGELOG.md
@@ -1,3 +1,14 @@
+# Change Log
+
+## [8.0 Version 1 Release 2] (2023-09-21)
+
+#### Release Notes
+- Updated content for V1R2 of the vSphere 8.0 vCenter,ESXi, and VM STIG Readiness Guides
+- ESXi SSH controls updated with new PowerCLI check and fix text
+- ESXI-80-000203,205,206,208 removed
+- VCSA-80-000009 removed. TLS configuration no longer supported. VCSA now only supports TLS 1.2
+- Misc. tech edits
+
 ## [8.0 Version 1 Release 1] (2023-04-18)
 
 #### Release Notes